backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
psycopg2-binary
pycryptodome==3.20.0
bcrypt==4.1.2
paho-mqtt==2.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
